Dish.java: change price type double to float

+++ src/main/java/seedu/duke/data/dish/Dish.java
@@ -5,9 +5,9 @@ import java.util.ArrayList;
 public class Dish {
     private String name;
     private ArrayList<Ingredient> ingredients;
-    private double price;
+    private float price;
 
-    public Dish(String name, ArrayList<Ingredient> ingredients, double price) {
+    public Dish(String name, ArrayList<Ingredient> ingredients, float price) {
         this.name = name;
         this.ingredients = ingredients;
         this.price = price;
